+++ src/tr/trVertex.hpp
@@ -9,15 +9,15 @@ namespace tr
 		        Vertex();
 		        Vertex(const Vector4& worldPosition, const Vector3& normal, const Vector2& textureCoord);
 		        Vertex(const Vector4& worldPosition, const Vector4& position, const Vector3& normal, const Vector2& textureCoord);
-		        Vertex(const Vector4& worldPosition, const Vector4& position, const Vector3& normal, const Vector2& textureCoord, const double inverseW);
+		        Vertex(const Vector4& worldPosition, const Vector4& position, const Vector3& normal, const Vector2& textureCoord, const float inverseW);
 
 		Vertex& operator+=(const Vertex& rhs);
 		Vertex& operator-=(const Vertex& rhs);
 
 		Vertex  operator+(const Vertex& rhs) const;
 		Vertex  operator-(const Vertex& rhs) const;
-		Vertex  operator*(const double& rhs) const;
-		Vertex  operator/(const double& rhs) const;
+		Vertex  operator*(const float& rhs) const;
+		Vertex  operator/(const float& rhs) const;
 
 		Vertex& normalize();
 
@@ -25,6 +25,6 @@ namespace tr
 		Vector4 position;
 		Vector3 normal;
 		Vector2 textureCoord;
-		double  inverseW;
+		float   inverseW;
 	};
 }
